ITT-576 - Display username in logout button
